+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [#49] Adds error screen for JSON RPC connection failure.
 - [#33] Exposes environment and chain information.
 - [#103] Adds instant validation feedback on input fields.
+- [#15] Adds a button to the header linking to a faucet on testnets.
 
 ## [0.8.0] - 2019-01-25
 ### Changed
@@ -67,6 +68,7 @@ token network.
 [#34]: https://github.com/raiden-network/webui/issues/34
 [#33]: https://github.com/raiden-network/webui/issues/33
 [#31]: https://github.com/raiden-network/webui/issues/31
+[#15]: https://github.com/raiden-network/webui/issues/15
 [#13]: https://github.com/raiden-network/webui/issues/13
 [#9]: https://github.com/raiden-network/webui/issues/9
 [#8]: https://github.com/raiden-network/webui/issues/8